+++ src/jvm/ml/topology/weka/OnlineEnsembleClassifierTopology.java
@@ -60,7 +60,7 @@ public class OnlineEnsembleClassifierTopology extends EnsembleLearnerTopologyBas
                 OnlineBinaryClassifierFactory(WekaOnlineClassificationAlgorithms.stochasticGradientDescent.name(), windowSize);
         final QueryFunction metaQueryFunction = new BinaryClassifierQuery.MetaQuery();
         final ReducerAggregator drpcPartitionResultAggregator =  new EnsembleLabelDistributionPairAggregator();
-        final QueryFunction<MlStormWekaState, Map.Entry<Double, double[]>> queryFunction = new BinaryClassifierQuery();
+        final QueryFunction<MlStormWekaState, Map.Entry<Integer, double[]>> queryFunction = new BinaryClassifierQuery();
 
         final List<StateUpdater> stateUpdaters = new ArrayList<StateUpdater>();
         final List<StateFactory> factories = new ArrayList<StateFactory>();
